+++ codraft/config.py
@@ -92,6 +92,7 @@ class ViewSection(conf.Section, metaclass=conf.SectionMeta):
     sig_format = conf.Option()
     ima_format = conf.Option()
     show_label = conf.Option()
+    show_contrast = conf.Option()
 
     # If True, images are shown with the same LUT range as the first selected image
     ima_ref_lut_range = conf.Option()